+++ src/tests/packet_conv/Test_audio.py
@@ -1,5 +1,5 @@
 # coding: UTF-8
-"""Tests for ``packet_conv.py``
+"""Tests for ``packet_conv/audio.py``
 
 Test cases:
   * Can convert audio data in ``ndarray`` with[out] custom bytes to bytes, and reverse convert
@@ -17,7 +17,7 @@ Author:
 # For import top layer module
 import sys
 from pathlib import Path
-sys.path.append(Path(__file__).absolute().parent.parent.__str__())
+sys.path.append(Path(__file__).absolute().parent.parent.parent.__str__())
 
 
 import unittest
@@ -25,50 +25,50 @@ import unittest
 import numpy as np
 from pydub import AudioSegment
 
-from modules import packet_conv
+from modules.packet_conv import audio
 import AUDIO_PARAM
 
 
-class Test_packet_conv(unittest.TestCase):
-  def test_packet_enc_dec_ndarray_ext(self):
+class Test_packet_conv_audio(unittest.TestCase):
+  def test_audio_packet_enc_dec_ndarray_ext(self):
     aud_org = part_create_testdata_ndarray()
     ext_org = bytes([1, 2, 3, 4])
 
-    packet = packet_conv.packet_enc_ndarray(aud_org, ext_org)
-    aud_prc, ext_prc = packet_conv.packet_dec_ndarray(packet)
+    packet = audio.packet_enc_ndarray(aud_org, ext_org)
+    aud_prc, ext_prc = audio.packet_dec_ndarray(packet)
 
     self.assertTrue((aud_org == aud_prc).all())
     self.assertEqual(ext_org, ext_prc)
 
 
-  def test_packet_enc_dec_ndarray_noext(self):
+  def test_audio_packet_enc_dec_ndarray_noext(self):
     aud_org = part_create_testdata_ndarray()
     ext_org = bytes()
 
-    packet = packet_conv.packet_enc_ndarray(aud_org, ext_org)
-    aud_prc, ext_prc = packet_conv.packet_dec_ndarray(packet)
+    packet = audio.packet_enc_ndarray(aud_org, ext_org)
+    aud_prc, ext_prc = audio.packet_dec_ndarray(packet)
 
     self.assertTrue((aud_org == aud_prc).all())
     self.assertEqual(ext_org, ext_prc)
 
 
-  def test_packet_enc_dec_audioseg_ext(self):
+  def test_audio_packet_enc_dec_audioseg_ext(self):
     aud_org = part_create_testdata_audioseg()
     ext_org = bytes([1, 2, 3, 4])
 
-    packet = packet_conv.packet_enc_audioseg(aud_org, ext_org)
-    aud_prc, ext_prc = packet_conv.packet_dec_audioseg(packet)
+    packet = audio.packet_enc_audioseg(aud_org, ext_org)
+    aud_prc, ext_prc = audio.packet_dec_audioseg(packet)
 
     self.assertEqual(aud_org, aud_prc)
     self.assertEqual(ext_org, ext_prc)
 
 
-  def test_packet_enc_dec_audioseg_noext(self):
+  def test_audio_packet_enc_dec_audioseg_noext(self):
     aud_org = part_create_testdata_audioseg()
     ext_org = bytes()
 
-    packet = packet_conv.packet_enc_audioseg(aud_org, ext_org)
-    aud_prc, ext_prc = packet_conv.packet_dec_audioseg(packet)
+    packet = audio.packet_enc_audioseg(aud_org, ext_org)
+    aud_prc, ext_prc = audio.packet_dec_audioseg(packet)
 
     self.assertEqual(aud_org, aud_prc)
     self.assertEqual(ext_org, ext_prc)
